check if file in server is zip or mp4

+++ lib-content-scrapers/src/main/java/com/ustadmobile/lib/contentscrapers/ContentScraperUtil.java
@@ -1309,8 +1309,10 @@ public class ContentScraperUtil {
 
 
     public static void deleteFile(File content) {
-        if (!content.delete()) {
-            UMLogUtil.logError("Could not delete: " + content.getPath());
+        if (content != null) {
+            if (!content.delete()) {
+                UMLogUtil.logError("Could not delete: " + content.getPath());
+            }
         }
     }
 }
